+++ log4cpp.sh
@@ -7,6 +7,10 @@ build_requires:
   - autotools
 env:
   LOG4_ROOT: "$LOG4CPP_ROOT"
+prefer_system_check: |
+  ls $LOG4CPP_ROOT/include/ > /dev/null && \
+  ls $LOG4CPP_ROOT/lib/ > /dev/null && \
+  true
 ---
 #!/bin/bash -ex
 rsync -a $SOURCEDIR/* .